added api_call, data fetched, and hero names json file
API_call notebook defined a fetch data function that collects 100000 recent matches with all the information we need. The data is cleaned to include the variables we need, including match_id, radiant_win (representing the winner; if radiant_win = False then the dire team wins), avg_mmr(a quantitative measure of skill levels of players in that match. We can exclude possible skill level difference that may result in outliers or noise with this variable), and the heroes picked by the two teams.  
The heroes picked are represented by numbers, and the corresponding hero names can be found in the hero_names.json file, which is collected from the official documentation of OpenDota API.